Prevent storing duplicate RF/IT signals (#717)
* Check if signal is duplicate

* Fix storing duplicate RF/IR Codes

* Lint formatting

+++ main/ZgatewayRF.ino
@@ -94,7 +94,8 @@ void RFtoMQTT() {
       RFtoMQTTdiscovery(MQTTvalue);
 #  endif
       pub(subjectRFtoMQTT, RFdata);
-      Log.trace(F("Store val: %lu" CR), MQTTvalue);
+      // Casting "receivedSignal[o].value" to (unsigned long) because ArduinoLog doesn't support uint64_t for ESP's
+      Log.trace(F("Store val: %u" CR), (unsigned long)MQTTvalue);
       storeSignalValue(MQTTvalue);
       if (repeatRFwMQTT) {
         Log.trace(F("Pub RF for rpt" CR));
